backend/requirements.txt
Django
djangorestframework
drf-yasg
channels
requests
Pillow
bs4
celery
redis
channels_redis
django-filter
easy-thumbnails
django-image-cropping
pyjwt
cent
wheel
pytils
psycopg2-binary
#pygraphviz
django-extensions
django-schema-graph
django-modeltranslation
django-rosetta